Optional Nav Bar/Orbit Slider Included
I have started to develop a new Nav Bar without any of my own CSS. The only links that are functional are the "Chocolate" links in the drop down menu. This is an example of how the anchors will work on that scrolling site. There is also a functional orbit slider at the very bottom of the site.